Explicit cast to byte, this will likely need to be done anywhere we send a byte that is not a byte explicitly.

+++ Terraria/Terraria/NetMessage.cs
@@ -44,7 +44,7 @@ namespace Terraria
 				{
 					Player player = Main.player[number];
 					binaryWriter.Write((byte)number);
-					binaryWriter.Write(player.male ? 0 : 1);
+					binaryWriter.Write((byte)(player.male ? 0 : 1));
 					binaryWriter.Write((byte)player.hair);
 					binaryWriter.Write(text);
 					binaryWriter.Write(player.hairDye);
